Python ExpandedInfoGenerate Script
Created a python script for generating CSV files to be imported into the DescDB. The script requires a file path to a CSV file as its only argument. This CSV file is the sorted spreadsheet provided by Halal Metropolis. The 2 generated CSV files then need to be imported into the respectively named DB tables in phpMyAdmin.