Auto Scrolling and summary popup change
auto scrolling slider does not work. It triggers the idle mode after no input but the slider is not smoth and it overflows arcgis with too many requests.

Also changed the popup so the link to learn more is styled as a button and "ExpandedInfo.php?id=___" doesnt need to be typed in to every arcgis database entry.